% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_permission.R
\name{ask_permission}
\alias{ask_permission}
\title{ask_permission}
\usage{
ask_permission()
}
\value{
'logical' if TRUE, the user authorized saving in ExtData/Previous
}
\description{
Ask users for permission to write the previous options file.
}
