% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_bbox_from_file.R
\name{bbox_from_file}
\alias{bbox_from_file}
\title{bbox_from_file}
\usage{
bbox_from_file(file_path, out_crs)
}
\arguments{
\item{file_path}{The path of the spatial file.}

\item{out_crs}{The output projection (string format).}
}
\description{
Helper function used to retrieve the bounding box of a specified spatial file
recognised by GDAL or OGG: the function reads the extent using gdalinfo or ogrinfo,
converting it to a specified CRS.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}
