% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_proc_opts.R
\name{check_proc_opts}
\alias{check_proc_opts}
\title{check_proc_opts}
\usage{
check_proc_opts(proc_opts)
}
\arguments{
\item{proc_opts}{data frame of parameters passed by \code{MODIStsp}}
}
\value{
NULL - processing interrupted if any condition is not met
}
\description{
helper function used to check consistency of processing
parameters
}
