% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_usearia.R
\name{check_usearia}
\alias{check_usearia}
\title{check_usearia}
\usage{
check_usearia(use_aria, gui, mess)
}
\arguments{
\item{use_aria}{\code{logical} If TRUE, aria2c is used to accelerate download
(if available !).}

\item{gui}{\code{logical} Indicates if processing was called starting from an
interactive environment or not. If FALSE, processing messages are sent to a
log file instead than to the console, and gWidgets messages are suppressed}

\item{mess}{pointer to the message window used to show processing messages
when gui = TRUE}
}
\value{
use_aria \code{logical} Indicating if aria2c should be used to accelerate
download
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
