% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earthdata_token.R
\name{get_earthdata_token}
\alias{get_earthdata_token}
\title{Get an Earthdata Login bearer token from username/password}
\usage{
get_earthdata_token(user, password)
}
\arguments{
\item{user}{\code{character} Earthdata Login username.}

\item{password}{\code{character} Earthdata Login password.}
}
\value{
\code{character} a bearer access token.
}
\description{
Internal function that fetches (or, if none exists, creates) an
Earthdata Login user token via the URS API, using HTTP Basic authentication
with the user's Earthdata credentials. The returned bearer token is then
used to authenticate downloads from the NASA Earthdata Cloud.
}
\author{
Pasi Autio (2024)
}
\note{
License: GPL 3.0
}
\keyword{internal}
