% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_dates.R
\name{get_mod_dates}
\alias{get_mod_dates}
\title{Find MODIS dates included in selected processing period}
\usage{
get_mod_dates(dates, date_dirs)
}
\arguments{
\item{dates}{2- element string array specifying start/end dates (yyyy.mm.dd)
for which the http addresses of folders in lpdaac should be retrieved
(e.g., c("2015.1.1", "2015.12.31)}

\item{date_dirs}{data frame full list of folders in lpdaac archive for product of interest}
}
\value{
array of folder names containing data for the MODIS product acquired in
the period specified by "dates"
}
\description{
Accessory function to find the folders corresponding to the requested
dates period within the full list retrieved by get_moddirs
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2016) \email{luigi@ranghetti.info}

Lorenzo Busetto, phD (2017)
}
