% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_filenames.R
\name{get_mod_filenames}
\alias{get_mod_filenames}
\title{Find the names of MODIS images corresponding to the selected dates}
\usage{
get_mod_filenames(
  http,
  used_server,
  n_retries,
  date_dir,
  v,
  h,
  tiled,
  out_folder_mod,
  gui
)
}
\arguments{
\item{http}{\code{character} LP DAAC-style product URL carried in the product
options DB. Only used here to derive the product \code{short_name}/\code{version}.}

\item{used_server}{\code{character} can assume values "http" or "offline".}

\item{n_retries}{\code{numeric} number of times the access to the CMR server
should be retried in case of error before quitting, Default: 20.}

\item{date_dir}{\code{character} folder name (a.k.a. acquisition date, format
"YYYY.MM.DD") for which MODIS files are to be identified.}

\item{v}{\verb{integer array} containing a sequence of the vertical tiles of interest
(e.g., c(18,19)).}

\item{h}{\verb{integer array} containing a sequence of the horizontal tiles of interest
(e.g., c(3,4)).}

\item{tiled}{\code{numeric [0/1]} indicates if the product to be downloaded is
tiled or not tiled. 1 = tiled product; 0 = non-tiled product (resolution 0.05 deg).}

\item{out_folder_mod}{\code{character} folder where hdf files have to be stored.}

\item{gui}{\code{logical} indicates if processing was called within the GUI environment
or not. If not, processing messages are redirected direct to the log file.}
}
\value{
\verb{character array} containing names of HDF images corresponding to the
requested tiles available for the product in the selected date. On "http"
the returned vector carries an attribute \code{"urls"}, a named character vector
mapping each HDF file name to its Earthdata Cloud download URL.
}
\description{
Accessory function to find the names of HDF images corresponding
to a given date and interval of spatial tiles. Names (and their Earthdata
Cloud download URLs) are retrieved by querying the NASA Common Metadata
Repository (CMR).
}
\note{
License: GPL 3.0
}
\author{
Original code by Babak Naimi (\code{.getModisList}, in
\href{https://r-gis.net/?q=ModisDownload}{ModisDownload.R})
modified to adapt it to MODIStsp scheme and to http archive (instead than old
FTP) by:

Lorenzo Busetto, phD (2014-2016)

Luigi Ranghetti, phD (2016)
}
