% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reqbands.R
\name{get_reqbands}
\alias{get_reqbands}
\title{get_reqbands}
\usage{
get_reqbands(bands_indexes_matrix, indexes_bandsel, indexes_bandnames,
  quality_bandsel, quality_bandnames, out_prod_folder, file_prefix, yy, DOY,
  out_format, reprocess)
}
\arguments{
\item{bands_indexes_matrix}{PARAM_DESCRIPTION}

\item{indexes_bandsel}{\code{integer 0/1 array} array of length equal to the number
of Spectral Indexes available for the product (standard + user-provided),
set to  1 for indexes to be processed.}

\item{indexes_bandnames}{\code{character array} Abbreviated Names of SIs available
for the selected product (used to build output file names of SIs).}

\item{quality_bandsel}{\code{integer 0|1 array} array of length equal to number
of available QIs, set to  1 for indexes to be processed.}

\item{quality_bandnames}{\code{character array} Abbreviated Names of Quality
Indicators available for the selected product (used to build output file
names of QIs).}

\item{out_prod_folder}{PARAM_DESCRIPTION}

\item{file_prefix}{PARAM_DESCRIPTION}

\item{yy}{PARAM_DESCRIPTION}

\item{DOY}{PARAM_DESCRIPTION}

\item{out_format}{\code{character ["ENVI" | "GTiff"]} Desired output format.}

\item{reprocess}{\code{character ["Yes" | "No"]} If Yes, reprocess data for already
existing dates.}
}
\value{
req_bands_indexes
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
