% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yeardates.R
\name{get_yeardates}
\alias{get_yeardates}
\title{identify dates to be processed for a year}
\usage{
get_yeardates(download_range, yy, start_year, end_year, start_date, end_date)
}
\arguments{
\item{download_range}{\code{character ["full" | "seasonal"]} If "full", all the
available images between the starting and the ending dates are downloaded;
If "seasonal", only the images included in the season are downloaded
(e.g: if the starting date is 2005-12-01 and the ending is 2010-02-31, only
the images of December, January and February from 2005 to 2010 - excluding
2005-01, 2005-02 and 2010-12 - are downloaded).}

\item{yy}{year for which the processing dates need to be identified}

\item{start_date}{\code{character} Start date for images download and preprocessing
(yyyy.mm.dd).}

\item{end_date}{\code{character} End date for images download and preprocessing
(yyyy.mm.dd).}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper function needed to identify the ranges of dates to
be processed for a given year as a function of \code{download_range} selection
and starting/ending dates and years
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
