% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{gui_load_options}
\alias{gui_load_options}
\title{gui_load_options}
\usage{
gui_load_options(choice)
}
\description{
GUI Helper function used to load options from JSON and set
values of the GUI accordingly
}
