% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp_install_launcher.R
\name{install_MODIStsp_launcher}
\alias{install_MODIStsp_launcher}
\title{install_MODIStsp_launcher}
\usage{
install_MODIStsp_launcher(desktop_path = NA, bin_path = NA, sudo = FALSE,
  desktop_shortcut = FALSE)
}
\arguments{
\item{desktop_path}{path to the desktop entry (Linux; default: /usr/share/applications/MODIStsp.desktop) or to the link in the Start Menu
(Windows; default: Start Menu -> Programs -> MODIStsp). Note that the patch must contain also the file name.}

\item{bin_path}{(Linux only) path of the link to the bash script (default: /usr/bin/MODIStsp); a path included in the PATH environment variable
is suggested.}

\item{sudo}{(Linux only) logical value which indicates if the creation of bin_path and desktop_path requires root permissions; in this case,
the root password is requested when launching the function.}

\item{desktop_shortcut}{(Windows only) logical value which indicates if also a desktop shortcut should be created.}
}
\description{
Function which allows to use MODIStsp in batch mode by creating links
}
\details{
MODIStsp tool can be used also as a standalone tool my launching a bash/batch script, which is stored in the function files.
 In order to simply retrieve it, this function will create a desktop entry and a symbolic link to the bash script (in Linux)
 or a link in the Start Menu to the batch script (in Windows).
}
\examples{
# Linux: common installation (script in /usr/bin,
# desktop entry in /usr/share/applications)
# (requires administrator permissions)
\dontrun{
install_MODIStsp_launcher(sudo = TRUE)
  # the administrator password is asked interactively}

# Linux: installation in a directory which does
# not require administrator permissions
\dontrun{
install_MODIStsp_launcher(desktop_path = "~/Desktop/MODIStsp.desktop",
  bin_path = "~/bin/MODIStsp")}

# Windows: common installation
# (script in the Start Menu and shortcut on the desktop)
\dontrun{
install_MODIStsp_launcher(desktop_shortcut = TRUE)
  # FIXME it does not run due to symlink problems in Windows}
}
\author{
Luigi Ranghetti, phD (2015)
email: ranghetti.l@irea.cnr.it

license GPL 3.0
}

