% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_opts.R
\name{load_opts}
\alias{load_opts}
\title{Load MODIStsp processing options}
\usage{
load_opts(opts_jsfile)
}
\arguments{
\item{opts_jsfile}{Expected file name of the JSON file containing
processing options}
}
\value{
\code{data frame} general_opts, containing the processing options
retrieved from the JSON file (or the defaults set at first execution).
See also \code{MODIStsp_GUI} and \code{MODIStsp_process}
}
\description{
Load MODIStsp processing option from \code{opts_jsfile} if
it already exist, otherwise initialize processing options to default and
save them to \code{opts_jsfile}. Send warnings if options file is from an
old version
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}

Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}
