% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prodopts.R
\name{load_prodopts}
\alias{load_prodopts}
\title{Load characteristics of the different MODIS products}
\usage{
load_prodopts(gui)
}
\arguments{
\item{gui}{\code{logical} if TRUE, the function was called from an interactive
MODIStsp session.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Load characteristics of the different MODIS products from \code{prodopts_file}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
