% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_check_files.R
\name{MODIStsp_check_files}
\alias{MODIStsp_check_files}
\title{MODIStsp_check_files}
\usage{
MODIStsp_check_files(out_prod_folder, file_prefix, bandnames,
  bandsel_orig_choice, yy, DOY, out_format, indexes_bandnames, indexes_bandsel,
  quality_bandnames, quality_bandsel)
}
\arguments{
\item{out_prod_folder}{string output folder name}

\item{file_prefix}{string file prefix of the product (e.g., MOD13Q1)}

\item{bandnames}{string array Bandnames of the MODIS product}

\item{bandsel_orig_choice}{0/1 array Indicates which original modis layers were selected to be processed}

\item{yy}{string year}

\item{DOY}{string doy}

\item{out_format}{string Gtiff or ENVI}

\item{indexes_bandnames}{string array Name of available spectral indexes forthe product}

\item{indexes_bandsel}{0/1 array Indicates which ospectral indexes were selected to be processed}

\item{quality_bandnames}{string array Name of available Quality Indicators for the product}

\item{quality_bandsel}{0/1 array Indicates which Quality Indicators were selected to be processed}
}
\value{
check - logical = 1 if all expected output files are already existing
}
\description{
Accessory function used to see if all expected out files for the selected date are already present.
}
\details{
if all expected out files for the selected date are already present, check_files is set to TRUE, and date is skipped
in MODIStsp_process
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}

Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}

