% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_download.R
\name{MODIStsp_download}
\alias{MODIStsp_download}
\title{MODIStsp download function}
\usage{
MODIStsp_download(
  modislist,
  out_folder_mod,
  download_server,
  urls,
  n_retries,
  use_aria,
  date_dir,
  year,
  DOY,
  token,
  sens_sel,
  date_name,
  gui,
  verbose
)
}
\arguments{
\item{modislist}{\verb{character array} List of MODIS images to be downloaded for
the selected date (as returned from \code{get_mod_filenames}). Can be a single
image, or a list of images in case different tiles are needed!}

\item{out_folder_mod}{\code{character} Folder where the hdfs are to be stored}

\item{download_server}{\code{character ["http"]} Server to be used.}

\item{urls}{\verb{named character} mapping each HDF file name in \code{modislist} to its
Earthdata Cloud download URL (the \code{"urls"} attribute returned by
\code{get_mod_filenames}).}

\item{n_retries}{\code{numeric} Max number of retry attempts on download. If
download fails more that n_retries times consecutively, abort}

\item{use_aria}{\code{logical} if TRUE, download using aria2c}

\item{date_dir}{\verb{character array} Sub-folder where the different images
can be found (element of the list returned from \code{get_mod_dirs}).}

\item{year}{\code{character} Acquisition year of the images to be downloaded}

\item{DOY}{\verb{character array} Acquisition doys of the images to be downloaded}

\item{token}{\code{character} Earthdata Login bearer token used to authenticate
downloads from the NASA Earthdata Cloud.}

\item{sens_sel}{\code{character ["terra" | "aqua"]} Selected sensor.}

\item{date_name}{\code{character} Date of acquisition of the images to be downloaded.}

\item{gui}{\code{logical} Indicates if on an interactive or non-interactive execution
(only influences where the log messages are sent).}

\item{verbose}{\code{logical} If FALSE, suppress processing messages, Default: TRUE}
}
\value{
The function is called for its side effects
}
\description{
Internal function dealing with download of MODIS hdfs from
the NASA Earthdata Cloud for a given date. Granules are downloaded from the
URLs discovered through the CMR (see \code{get_mod_filenames}), authenticating
with an Earthdata Login bearer token.
}
\author{
Lorenzo Busetto, phD (2014-2017)

Luigi Ranghetti, phD (2015)
}
