% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_extract.R
\name{MODIStsp_extract}
\alias{MODIStsp_extract}
\title{MODIStsp_extract}
\usage{
MODIStsp_extract(in_rts, start_date, end_date, sp_object, FUN = "mean",
  na.rm = T, verbose = F)
}
\arguments{
\item{in_rts}{"rasterStackTS" or "rasterBrickTS" object created by MODIStsp}

\item{start_date}{"Date", "POSIXct" or "POSIXlt" starting date of the period to be considered for data extraction}

\item{end_date}{"Date", "POSIXct" or "POSIXlt" ending date of the period to be considered for data extraction}

\item{sp_object}{"sp" object specifying the "positions" from which data has to be extracted
   If sp_object represents lines, the function returns the values of the cells of a in_rts object that are touched by the line.
If y represents polygons, the function  returns the values of the cells of a in_rts object that are covered by a polygon. A cell
is covered if its center is inside the polygon .}

\item{FUN}{function to summarize the values (e.g. mean) on polygon data frames. The function should take a single numeric vector as argument and return a
single value (e.g. mean, min or max), and accept a na.rm argument. Thus, standard R functions not including an na.rm argument must
be wrapped as in this example: fun=function(x,...)length(x).}

\item{na.rm}{Logical If TRUE, and sp_object is a polygon, then na.rm = T is used when applying the function
to the different pixels of the polygon. Default = T.}

\item{verbose}{Logical If TRUE, messages on extraction completion are sent out. Default = T.}
}
\value{
xts object. Each column of data corresponds to one poin or one polygon
}
\description{
function used to extract time series data from rts files created by MODIStsp
on spatial locations provided in the form of "R" spatial objects (spatialPoints,
spatialPolygons, etc.)
}
\details{
The function takes as input an rasterStackTS object created by MODIStsp, a starting and ending date
and a standard "R" spatial object, and provides as time series for the spatial locations
specified in the spatial object in the form of a "R" xts object .
If the input spatial object is a "point" or "line" one, the xts output object contains one column
for each specified point, or for each cell intersecting the line, and one line for each date.
If the input spatial object is a "polygon" one, the xts output object contains one column
for each polygon, containing values obtained applying the function specified as the FUN
argument over all pixels belonging to the polygon, and one line for each date.
}
\examples{
# Extract average and standard deviation values from a rts object created by MODIStsp
# for each polygon of a shapefile, for each date in the period between 2001-01-01 and 2014-12-31
\dontrun{
#Set the inputs
infile = "in_path/MOD13Q1_MYD13Q1_NDVI_49_2000_353_2015_RData.RData"  # Input rts file
shpname = "path_to_file/rois.shp"  # Path to Polygon Shapefile
startdate = as.Date("2010-01-01")  # Start date for extraction
enddate = as.Date("2014-12-31")    # End date for extraction
#Load Data
inrts = get(load(infile))          # Load the rts file
shape = readOGR(dirname(shpname), 'rois') # Read shapefile to SpatialPolygons
# Compute average and St.dev
dataavg = MODIStsp_extract(inrts, startdate, enddate, shape, FUN = 'mean', na.rm = T)
datasd = MODIStsp_extract(inrts, startdate, enddate, shape, FUN = 'sd', na.rm = T)
plot(dataavg)
}
}
\author{
Lorenzo Busetto, phD (2015)
email: busetto.l@irea.cnr.it
}

