% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_get_prodlayers.R
\name{MODIStsp_get_prodnames}
\alias{MODIStsp_get_prodnames}
\title{Retrieve the names of all available product}
\usage{
MODIStsp_get_prodnames()
}
\value{
character array of product names
}
\description{
Function used to retrieve the names of available MODIS products
}
\note{
License: GPL 3.0
}
\examples{
# Get MODIStsp product names
MODIStsp_get_prodnames()
}
\author{
Lorenzo Busetto, phD (2014-2020) \email{lbusett@gmail.com}
}
