% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp_GUI.R
\name{MODIStsp_GUI}
\alias{MODIStsp_GUI}
\title{MODIStsp_GUI}
\usage{
MODIStsp_GUI(general_opts)
}
\arguments{
\item{general_opts}{General options data frame passed by MODIStsp_main. Contains paths and other variables used to initialize the GUI
if a previous options file is not existing.}
}
\value{
NULL - Processing options are saved in "previous" file and (if "Save options" is pressed) in user's selected file
}
\description{
Function used to generate and handle the GUI used to allow selection of MODIStsp processing parameters
 If the "previous options" file (MODIStsp_Previous.RData) already exists, it is loaded and used to reinstate
	the GUI to its last state. Otherwise, the previous options file is created by launching the MODIStsp_read_xml fucntion
}
\author{
Lorenzo Busetto, phD (2014)
email: busetto.l@irea.cnr.it

license GPL(>2)
}

