% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp_set_opts.R
\name{MODIStsp_set_opts}
\alias{MODIStsp_set_opts}
\title{MODIStsp_set_opts OBSOLETE ! CHANGED TO READING FROM XML FILE ! TO BE REMOVED FROM PROJECCT}
\usage{
MODIStsp_set_opts(previous_file = previous_file)
}
\arguments{
\item{previous_file}{previous_file Name of file used intrnally by MODIStsp project to store previously used processing
options}
}
\description{
Function used to define the structure of the "opts" variable used to store all processing options for the different
products and generic, and save it in the "previous" file.
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014) \email{busetto.l@irea.cnr.it}
}

