% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_vrt_create.R
\name{MODIStsp_vrt_create}
\alias{MODIStsp_vrt_create}
\title{Create MODIStsp virtual files}
\usage{
MODIStsp_vrt_create(sensor, out_prod_folder, bandnames, indexes_bandnames,
  quality_bandnames, bandsel, indexes_bandsel, quality_bandsel, file_prefixes,
  sens_sel, ts_format, nodata_value, out_format, rts)
}
\arguments{
\item{out_prod_folder}{\code{character} Main output folder.}

\item{file_prefixes}{\code{character array (2)} file_prefixes for TERRA and AQUA -
used to identify the files corresponding to each sensor}

\item{sens_sel}{\code{character ["Aqua", "Terra", "Mixed"]} Sensor for which the
virtual files are to be created. If "Mixed" and both TERRA and AQUA
images are available, a "mixed" virtual file comprising data from both sensors
ordered on DOY base is also created.}

\item{ts_format}{\code{character ["ENVI" | "GDAL" | "Both"]} Required output format
for virtual file.}

\item{nodata_value}{\code{character} NoData value to be used for vrt files
(equal to NoData value of inputs)}

\item{out_format}{\code{character ["ENVI" | "GTiff"]} Format of images used as
"input" for the vrt and contained in out_prod_folder/band folders.}

\item{rts}{\code{character ["Yes" | "No"]} If Yes, create "R" RasterStacks time series}

\item{meta_band}{\code{character} Name of the band (or SI, or QI) for which the
virtual file has to be created.}
}
\value{
NULL -
}
\description{
Function used to create virtual files from time series of single-band
files corresponding to different acquisition dates. The function takes as input
the folder in which the single-band files are stored, and creates a ENVI Meta
file and/or a GDAL vrt file that allows access to the full time series as if
it was a single physical file.
Created virtual files are stored in the "Time Series" subfolder of `out_prod_folder``
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}

Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}
