% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_message.R
\name{process_message}
\alias{process_message}
\title{process_message}
\usage{
process_message(mess_text, gui, mess_lab)
}
\arguments{
\item{mess_text}{PARAM_DESCRIPTION}

\item{gui}{PARAM_DESCRIPTION}

\item{mess_lab}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper function to provide processing messages
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
