% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_helpers.R
\name{update_tiles}
\alias{update_tiles}
\title{update_tiles}
\usage{
update_tiles(bbox_out, output_proj4_wid, mod_proj_str, modis_grid, start_x_wid,
  end_x_wid, start_y_wid, end_y_wid)
}
\description{
Helper function to update the selected tiles with the intersection with ####
the bounding box
}
