% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_from_file.R
\name{bbox_from_file}
\alias{bbox_from_file}
\title{Retrieve bbox from a spatial file}
\usage{
bbox_from_file(file_path, crs_out)
}
\arguments{
\item{file_path}{\code{string} path of a spatial file.}

\item{crs_out}{\code{string} proj4string of the desired output projection.}
}
\description{
Helper function used to retrieve the bounding box of a specified spatial file
recognized by GDAL or OGG: the function reads the extent using gdalinfo or ogrinfo,
converting it to a specified CRS.
}
\note{
License: GPL 3.0
}
\author{
Luigi Ranghetti, phD (2015-2017) \email{ranghetti.l@irea.cnr.it}
}
