% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_from_file.R
\name{bbox_from_file}
\alias{bbox_from_file}
\title{Retrieve bbox from a spatial file}
\usage{
bbox_from_file(file_path, crs_out)
}
\arguments{
\item{file_path}{\code{character} path of a spatial file.}

\item{crs_out}{(\code{crs} | \code{character}) crs of the desired output projection,
or string coercible to it using \code{sf::st_crs()} (e.g., WKT or numeric
EPSG code)}
}
\description{
Helper function used to retrieve the bounding box of a specified spatial file
recognized by  \code{sf} or \code{raster}: the function reads the extent using \code{sf::st_bbox()}
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2017)

Luigi Ranghetti, phD (2017)
}
