% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earthdata_token.R
\name{get_earthdata_token}
\alias{get_earthdata_token}
\title{Get Earthdata Bearer access token function}
\usage{
get_earthdata_token(user, password)
}
\arguments{
\item{user}{\code{character} Username for Earthdata servers}

\item{password}{\code{character} Password for Earthdata servers}
}
\value{
The function is called for its side effects
}
\description{
Internal function to fetch Earthdata access token
}
\details{
The function is used to:
\itemize{
\item Fetch Bearer token if there is one already defined;
\item Request new token if no token is defined
\item Update token if the token is expired
\item Bearer token is used for authentication by other functions
}
}
