% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prodopts.R
\name{load_prodopts}
\alias{load_prodopts}
\title{Load characteristics of the different MODIS products}
\usage{
load_prodopts()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Load characteristics of the different MODIS products from \code{prodopts_file}
}
\author{
Lorenzo Busetto, phD (2017)
}
