% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_lpdaac_accessoires.R
\name{lpdaac_getmod_dirs}
\alias{lpdaac_getmod_dirs}
\title{lpdaac_getmod_dirs}
\usage{
lpdaac_getmod_dirs(ftp, http, used_server = NA, user = user,
  password = password, gui, out_folder_mod, .Platform)
}
\arguments{
\item{ftp}{string ftp site corresponding to a given MODIS product}

\item{http}{string http site on lpdaac corresponding to a given MODIS product}

\item{used_server}{string can assume values "http" or "ftp" depending on the used download server; if NA, the script tries to download with http, using ftp if the download fails}

\item{user}{username for earthdata server}

\item{password}{password for earthdata server}

\item{gui}{logical indicates if processing was called within the GUI environment or not. If not, direct processing messages to the log}

\item{out_folder_mod}{ouput folder for original HDF storage}

\item{.Platform}{string os platform (from call to .Platform)}
}
\value{
list of all available folders (a.k.a. dates) for the requested MODIS product on lpdaac archive
}
\description{
Accessory function to get the full list of directories on the lpdaac http site (modified after Barry Rowlingson function):
}
\note{
License: GPL 3.0
}
\author{
Original code by Babak Naimi (.getModisList, in ModisDownload.R - http://r-gis.net/?q=ModisDownload )
Modified to adapt it to MODIStsp scheme and to http archive (instead than old FTP) by Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}
}
