% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp_lpdaac_accessoires.R
\name{lpdaac_getmod_dirs}
\alias{lpdaac_getmod_dirs}
\title{lpdaac_getmod_dirs}
\usage{
lpdaac_getmod_dirs(http, .Platform)
}
\arguments{
\item{http}{string http site on lpdaac corresponding to a given MODIS product}

\item{.Platform}{string os platform (from call to .Platform)}
}
\value{
list of all available folders (a.k.a. dates) for the requeted MODIS product on lpdaac archive
}
\description{
Accessory function to get the full list of directories on the lpdaac http site (modified after Barry Rowlingson function):
}
\note{
License: GPL 3.0
}
\author{
Original code by XXXXXX
Modified to adapt it to MODIStsp scheme and to http archive (instead than old FTP) by Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}
}

