% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp.R
\name{MODIStsp}
\alias{MODIStsp}
\title{MODIStsp_main Launch the tool}
\usage{
MODIStsp(gui = TRUE, options_file = NULL, MODIStsp_dir = NA)
}
\arguments{
\item{gui}{logical parameter (TRUE: the GUI is opened before processing; FALSE: the saved parameters are directly passed)}

\item{options_file}{settings (optional): full path of the RData file containing the processing options (default: Previous.RData in subdir Previous);}

\item{MODIStsp_dir}{main directory of the local installation of the tool(if not gived, it is automatically retrieved, but this works only if the function is launched from here!)}
}
\description{
Main function for the MODIS Time Series Processing Tool (MOD_TSP)
}
\details{
The function is used to initialize the processing (folder names, packages, etc.), to launch the GUI (MODIStsp_GUI) and receive its outputs,
 and to launch the required routines for downloading and processing the requested datasets.
}
\author{
Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it},
Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it},
license GPL 3.0
}

