% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp.R
\name{MODIStsp}
\alias{MODIStsp}
\title{Main MODIStsp function}
\usage{
MODIStsp(gui = TRUE, options_file = NULL, spatial_file_path = NULL,
  scroll_window = FALSE, test = -1)
}
\arguments{
\item{gui}{\code{logical} if TRUE: the GUI is opened before processing. If FALSE:
processing parameters are retrieved from the provided \code{options_file}
argument), Default: TRUE}

\item{options_file}{\code{character} full path to a JSON file
containing MODIStsp processing options saved from the GUI. If NULL,
parameters of the last succesfull run are retrieved from file
"MODIStsp_Previous.json" in subdir Previous), Default: NULL}

\item{spatial_file_path}{\code{character} (optional) full path of a spatial file
to use to derive te processing extent. If not NULL, the processing options
which define the extent, the selected tiles and the "Full Tile / Resized"
in the JSON options file are overwritten and new files are created on the
extent of the provided spatial file, Default: NULL}

\item{scroll_window}{\code{logical} if TRUE, the GUI window is opened
fullscreen with scrollbars (this is useful on devices with small displays).
If using a device with a display resolution >= 1024x768, leaving this
parameter to FALSE is suggested, Default: FALSE}

\item{test}{\code{integer} if set, MODIStsp is executed in "test mode", using a
preset Options File instead than opening the GUI or accepting the
\code{options_file} parameter. This allows both to check correct installation on
user's machines, and to implement unit testing. The number indicates which
tests to execute (six are available). If == 0, the test is selected
randomly. If == -1, MODIStsp is executed in normal mode. Default: -1}
}
\description{
Main function for the MODIS Time Series Processing Tool
(MODIStsp)
}
\details{
The function is used to:
\itemize{
\item initialize the processing (folder names, packages, etc.);
\item launch the GUI (\code{\link[=MODIStsp_GUI]{MODIStsp_GUI()}}) and receive its outputs on interactive
execution;
\item load the options file on non-interactive execution;
\item launch the routines for downloading and processing the requested datasets.
(\code{\link[=MODIStsp_process]{MODIStsp_process()}})
}
}
\note{
License: GPL 3.0
}
\examples{
# Running the tool without any option will start the GUI with the default or
# last used settings
\dontrun{
MODIStsp()}

# Run the tool using the settings previously saved in a specific options file
\dontrun{
MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json")}

# Run the tool using the settings previously saved in a specific option file
# and specifying the extent from a spatial file
\dontrun{
MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json",
  spatial_file_path = "X:/yourpath/yourspatialfile.shp" )}

# Run the tool in batch mode, using the settings previously saved in a
# specific options file and specifying each time the extent from a different
# spatial file (e.g., to perform the same processing on different extents)

\dontrun{
extent_list = list.files("X:/path/containing/some/shapefiles/", "\\\\.shp$")
for (single_shape in extent_list)
  MODIStsp(gui = FALSE, options_file = "X:/yourpath/youroptions.json",
    spatial_file_path = single_shape )}
}
\seealso{
\code{\link[=MODIStsp_GUI]{MODIStsp_GUI()}}, \code{\link[=MODIStsp_process]{MODIStsp_process()}}
}
\author{
Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}

Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}
