% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp_addindex.R
\name{MODIStsp_addindex}
\alias{MODIStsp_addindex}
\title{MODIStsp_addindex}
\usage{
MODIStsp_addindex()
}
\arguments{
\item{Previous_File}{string path to the MODIStsp_Previous.RData file}
}
\value{
NULL - the MODIStsp_Previous.RData file is modified so to allow computation of the additional index
}
\description{
Function used to add a user-defined Spectral Index to the default list of computable spectral indexes
}
\details{
The function asks the user to provide the info related to the new desired Spectral Index using a GUI interface,
checks for correctness of provided information (e.g., correct bandnames, computable formula, etc...). If the index is legit,
it modifies the MODIStsp_Previous.RData file so to allow computation of the additional index within MODIStsp.
To remove all custom-added spectral indexes, simply delete the MODIStsp_Previous.RData file within the /Extdata subfolder of the
folder in which the package was installed
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2015) \email{busetto.l@irea.cnr.it}
}

