% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_GUI.R
\name{MODIStsp_GUI}
\alias{MODIStsp_GUI}
\title{Build and manage the MODIStsp GUI}
\usage{
MODIStsp_GUI()
}
\value{
the function is called for its side effects - opening
the GUI and allowing to set, save, load options and eventually
launch the processing.
}
\description{
Function used to generate and handle the GUI used to allow selection of
MODIStsp processing parameters. If the "previous options" file
(MODIStsp_Previous.json) already exists, it is loaded and used to reinstate
the GUI to its last state. Otherwise, the previous options file is created
by launching the MODIStsp_read_xml function
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014-2017) \email{lbusett@gmail.com}

Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}
