% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_process.R
\name{MODIStsp_process}
\alias{MODIStsp_process}
\title{MODIStsp_process}
\usage{
MODIStsp_process(p_opts)
}
\arguments{
\item{p_opts$sel_prod}{string selected MODIS product}

\item{p_opts$start_date}{string p_opts$start_date for images download and preprocessing
(yyyy.mm.dd)}

\item{p_opts$end_date}{string p_opts$end_date for images download and preprocessing
(yyyy.mm.dd)}

\item{p_opts$out_folder}{main output folder}

\item{p_opts$out_folder_mod}{output folder for original HDF storage}

\item{reprocess}{string string ("Yes"/"No") If Yes, reprocess data for already existing
dates (Default = 'Yes')}

\item{p_opts$delete_hdf}{string ("Yes"/"No") If Yes, delete original HDF after completion}

\item{p_opts$sensor}{string ("Terra" or "Aqua" or "Both")}

\item{https}{hash https site for download of HDF of selected product}

\item{p_opts$ftps}{hash p_opts$ftps site for download of HDF of selected product}

\item{p_opts$download_server}{service used to download MODIS tiles, one of: 'http', 'ftp', NA.}

\item{p_opts$user}{p_opts$username for http download
(\href{https://urs.earthdata.nasa.gov/home}{urs.earthdata.nasa.gov/home})}

\item{p_opts$password}{p_opts$password for http download
(\href{https://urs.earthdata.nasa.gov/home}{urs.earthdata.nasa.gov/home})}

\item{p_opts$start_x}{int start horizontal tile}

\item{p_opts$start_y}{int start vertical tile}

\item{p_opts$end_x}{int end horizontal tile}

\item{p_opts$end_y}{int end vertical tile}

\item{p_opts$bbox}{array output bounding box (xmin, xmax, ymin, ymax ) in out proj coords}

\item{p_opts$out_format}{string output raster format (ENVI or GTiff)}

\item{p_opts$compress}{string p_opts$compression for GTiff outputs (None, LZW, DEFLATE)}

\item{p_opts$native_res_sel}{string "Native" or "Resampled"}

\item{p_opts$native_res}{float Output resolution (in output projection measurement unit)}

\item{native_res}{float Native resolution of MODIS product}

\item{p_opts$tiled}{0/1 1 = p_opts$tiled product; 0 = non-p_opts$tiled product (resolution 0.05 deg)}

\item{p_opts$mod_proj_str}{string proj4 string for MODIS product native projection}

\item{p_opts$outproj_str}{string proj4 string of selected output projection}

\item{p_opts$nodata_in}{array Original NoData for MODIS bands}

\item{p_opts$nodata_out}{Target NoData for MODIS bands}

\item{p_opts$nodata_out}{string (Yes/No) if Yes, NoData are set to p_opts$nodata_out in
output rasters}

\item{p_opts$scale_val}{string (Yes/No) if Yes, output values in are rescaled in the
measure unit of the variable}

\item{p_opts$rts}{string ("Yes"/"No") If Yes, create p_opts$rts time series}

\item{p_opts$datatype}{string array p_opts$datatypes of MODIS bands}

\item{bandsel}{array of length equal to number of original modis layers. set
to 1 for bands to be processed}

\item{p_opts$bandnames}{array of Abbreviated Names of MODIS bands}

\item{p_opts$indexes_bandsel}{array of length equal to number of available spectral p_opts$indexes,
set to  1 for p_opts$indexes to be processed}

\item{p_opts$indexes_bandnames}{array of Abbreviated Names of MODIS p_opts$indexes}

\item{p_opts$indexes_formula}{array of p_opts$indexes formulas}

\item{p_opts$indexes_nodata_out}{NoData values for p_opts$indexes}

\item{p_opts$quality_bandnames}{array of  Names of MODIS p_opts$quality indicators}

\item{p_opts$quality_bandsel}{array of length equal to number of available p_opts$quality indicators,
set to  1 for indicators to be processed}

\item{p_opts$quality_bitN}{list of strings with number of entries equal to number of
p_opts$quality indicators. each entry contains position of bits corresponding to a QI
(e.g., 0-1)}

\item{p_opts$quality_source}{list of strings which connects each p_opts$quality indicator to
its source aggregated p_opts$quality assurance layer}

\item{p_opts$quality_nodata_in}{Always 255}

\item{p_opts$full_ext}{string ("p_opts$full_ext" or "Resized")}

\item{p_opts$quality_nodata_out}{Always 255}

\item{p_opts$file_prefixes}{output file prefix according to selected product (e.g., MOD13Q1)}

\item{p_opts$main_out_folder}{Suffix to add to the overall p_opts$out_folder to create the out
dir for the product (corresponds to an abbreviation of the selected product)}

\item{p_opts$resampling}{string p_opts$resampling method (near, bilinear, etc.)}

\item{p_opts$ts_format}{string format of virtual files (None, ENVI Meta Files, GDAL vrt files,
ENVI and GDAL)}

\item{p_opts$gui}{logical indicates if processing was called within the p_opts$gui environment or not.
If not, direct processing messages to the log}

\item{p_opts$use_aria}{logical if TRUE, then aria2c is used to accelerate download
(if available !)}

\item{p_opts$download_range}{character if "full", all the available images between the
starting and the ending dates are downloaded; if "seasonal", only the images
included in the season (e.g: if the starting date is 2005-12-01 and the ending
is 2010-02-31, only the images of December, January and February from 2005 to 2010
\itemize{
\item excluding 2005-01, 2005-02 and 2010-12 - are downloaded)
}}

\item{p_opts$scale_factor}{scale factor to be applied to MODIS layer to convert from
scaled integer to correct measure units}

\item{p_opts$offset}{p_opts$offset to be applied to MODIS layer to convert form scaled integer
to correct measure units}
}
\description{
main function of MODIStsp tool. Takes as input processing parameters
specified by the p_opts$user using MODIStsp_p_opts$gui and saved in MODIStsp_Previous.json
(Interactive use), or a p_opts$user specified JSON file (batch use)
(See MODIStsp_main for details ) and performs all required
processing.
}
\details{
After retrieving the input processing options, the function
\enumerate{
\item Accesses lpdaac http or ftp archive to determine the list of dates to be processed
\item Performs all required processing steps on each date (download, reprojection,
resize, mosaicing, p_opts$indexes computation, p_opts$quality indicators computation)
\item Creates virtual files of the processed time series.
}

Reprojection and resize is dealt with by accessing gdal routines through the \code{gdaUtils}
package.
Extraction of bitfields from p_opts$quality layers is done though fast bitwise computation
Checks are done in order to not re-download already existing HDF images, and not
reprocess already processed dates (if the p_opts$user did not specify that)
}
\note{
Thanks Tomislav Hengl and Babak Naimi, whose scripts made the starting point for
development of this function (\href{http://r-gis.net/?q=ModisDownload}{ModisDownload};
\href{http://spatial-analyst.net/wiki/index.php?title=Download_and_p_opts$resampling_of_MODIS_images}{Download_and_p_opts$resampling_of_MODIS_images})

License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014-2015) \email{busetto.l@irea.cnr.it}

Luigi Ranghetti, phD (2015) \email{ranghetti.l@irea.cnr.it}
}
