% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_read_xml.R
\name{MODIStsp_read_xml}
\alias{MODIStsp_read_xml}
\title{Read MODIS products characteristics from XML}
\usage{
MODIStsp_read_xml(prodopts_file, xml_file)
}
\arguments{
\item{prodopts_file}{string filename of the RData in which to store the data
parsed from the XML file}

\item{xml_file}{string filename of the XML file containing the MODIS products
characteristics}
}
\value{
NULL - retrieved data are stored in the specified RData file
}
\description{
function used to parse the XML file used to store the characteristics of
MODIS Land Products and store them in the "prod_opts" data frame
}
\details{
The function parses the XML file product by product, stores data in a data frame
and saves the data frame within the "MODIStsp_previous" RData file as a list of lists
}
\note{
License: GPL 3.0
}
\author{
Lorenzo Busetto, phD (2014-2017)

Luigi Ranghetti, phD (2015)
}
