% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODIStsp_vrt_create.R
\name{MODIStsp_vrt_create}
\alias{MODIStsp_vrt_create}
\title{MODIStsp_vrt_create}
\usage{
MODIStsp_vrt_create(out_prod_folder, meta_band, file_prefixes, sens_sel,
  ts_format, nodata_value, out_format)
}
\arguments{
\item{out_prod_folder}{string main output folder}

\item{meta_band}{string "name" of the band (or index, or qi) for which the virtual file is to be created}

\item{file_prefixes}{string file_prefixes for terra and aqua - used to identify the files corresponding to each sensor}

\item{sens_sel}{string name of the sensor for which the time serie has to be created (Aqua, Terra, Mixed) If "Mixed" and both terra and aqua
images are available, a "mixed" virtual file comprising data from both sensors ordered on DOY base is created}

\item{ts_format}{string required output format for virtual file (ENVI, GDAL, Both)}

\item{nodata_value}{string nodata value to be used for vrt files (equal to nodata value of inputs)}

\item{out_format}{format of images used as "input" for the vrt and contained in out_prod_folder/band folders (ENVI or GTiff)}
}
\value{
NULL - virtual files are stored in the "Time Series" subfolder of out_prod_folder
}
\description{
Function used to create virtual files from time series of single-band files corresponding to
different acquisition dates
}
\details{
The function takes as input the folder in which the single-band files are stored, and cretes a ENVI Meta file and/or
a GDAL vrt file that allows access to the full time series as if it was a single physical file
}
\author{
Lorenzo Busetto, phD (2014-2015)
email: busetto.l@irea.cnr.it
Luigi Ranghetti, phD (2015)
license GPL 3.0
}

