% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_message.R
\name{process_message}
\alias{process_message}
\title{Spawn processing update messages}
\usage{
process_message(mess_text, gui, mess_lab, verbose = TRUE)
}
\arguments{
\item{mess_text}{\code{character} text to be shown in the processing windows
and/or the console}

\item{gui}{\code{logical} indicating if the message should be passed to the
status window or only to the console.}

\item{mess_lab}{pointer to the gwindow used to shoe messages when gui == TRUE}

\item{verbose}{\code{logical} If FALSE, suppress processing messages, Default: TRUE}
}
\value{
The function is called for its side effects
}
\description{
helper function to provide processing messages
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
