% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_message.R
\name{process_message}
\alias{process_message}
\title{Spawn processing update messages}
\usage{
process_message(mess_text, verbose = TRUE)
}
\arguments{
\item{mess_text}{\code{character} text to be shown in the processing windows
and/or the console}

\item{verbose}{\code{logical} If FALSE, suppress processing messages, Default: TRUE}
}
\value{
The function is called for its side effects
}
\description{
helper function to provide processing messages
}
\author{
Lorenzo Busetto, phD (2017)
}
